fix: address code review findings, switch review to manual dispatch
- Fix float32 precision regression: cast to float64 before np.dot
  (variance was clamped to 0.0 for near-zero-var layers)
- Fix histogram_counts list[int] -> list[float] type mismatch
- Add NaN/Inf finite guard in FixedRangeHistogramQuantiles.update()
- Use float64 dtype in StreamingGlobalAggregator.update() mean/var
- Add from __future__ import annotations to tests/__init__.py
- Strengthen variance guard test to compare against float64 oracle
- Switch opencode PR review to manual dispatch only (/oc comment trigger)

+++ src/weightlens/aggregators/streaming_global.py
@@ -45,8 +45,8 @@ class StreamingGlobalAggregator(GlobalAggregator):
         if value_count == 0:
             return
 
-        batch_mean = float(np.mean(values))
-        batch_variance = float(np.var(values, ddof=0))
+        batch_mean = float(np.mean(values, dtype=np.float64))
+        batch_variance = float(np.var(values, ddof=0, dtype=np.float64))
         if not math.isfinite(batch_mean) or not math.isfinite(batch_variance):
             logger.error("Non-finite value encountered in global aggregation.")
             raise ValueError("Non-finite value encountered in global aggregation.")
